Refactor service architecture for addon support
- Updated `.gitignore` to ignore local tooling caches.
- Changed namespaces in unit tests to reflect new structure.
- Removed specific service mocks in `CarJobProcessorServiceTests` and introduced mocks for `IAddonServiceStrategy`.
- Added `AddonTypeAttribute` for easier registration of addon services.
- Refactored `HandWaxAndShineService`, `InteriorCleanService`, and `TireShineService` to implement `IAddonServiceStrategy`.
- Modified `CarJobProcessorService` to dynamically process addon services.
- Updated `README.md` with guidelines for adding new addon services.
- Ensured compatibility in `CarWashProcessor.csproj` with new service structure.
- Created `IAddonServiceStrategy` interface for consistent addon service implementation.
- Registered new addon service strategies in `Program.cs`.
