+++ CarWashProcessor.UnitTests/Application/Strategies/Addon/InteriorCleanServiceTests.cs
@@ -1,57 +1,60 @@
 // Copyright (c) 2025 Car Wash Processor, All Rights Reserved
 
+using CarWashProcessor.Application.Strategies.Addons;
 using CarWashProcessor.Models;
-using CarWashProcessor.Services;
+
 using Microsoft.Extensions.Logging;
+
 using Moq;
+
 using System.Collections.Immutable;
 
-namespace CarWashProcessor.UnitTests.Services
+namespace CarWashProcessor.UnitTests.Application.Strategies.Addon
 {
     [TestClass]
-    public class HandWaxAndShineServiceTests
+    public class InteriorCleanServiceTests
     {
-        private Mock<ILogger<HandWaxAndShineService>>? _loggerMock = null;
-        private HandWaxAndShineService? _washService = null;
+        private Mock<ILogger<InteriorCleanService>>? _loggerMock = null;
+        private InteriorCleanService? _washService = null;
         private CarJob? _carJob = null;
 
         [TestInitialize]
         public void TestInit()
         {
-            _loggerMock = new Mock<ILogger<HandWaxAndShineService>>();
+            _loggerMock = new Mock<ILogger<InteriorCleanService>>();
             _carJob = new CarJob(8675309, ECarMake.Ford, EServiceWash.Awesome, new ImmutableArray<EServiceAddon>());
         }
 
         [TestMethod]
         public void Ctor_WhenLoggerIsNull_ThrowsArgumentNullException()
         {
-            var exception = Assert.ThrowsException<ArgumentNullException>(() => _washService = new HandWaxAndShineService(null!));
+            var exception = Assert.ThrowsException<ArgumentNullException>(() => _washService = new InteriorCleanService(null!));
             Assert.AreEqual("logger", exception.ParamName);
         }
 
         [TestMethod]
         public void Ctor_WhenAllArgumentsProvided_Succeeds()
         {
-            _washService = new HandWaxAndShineService(_loggerMock!.Object);
+            _washService = new InteriorCleanService(_loggerMock!.Object);
             Assert.IsTrue(_washService is not null);
         }
 
         [TestMethod]
-        public async Task HandWaxAndShineAsync_WhenCarJobIsNull_ThrowsArgumentNullException()
+        public async Task CleanInteriorAsync_WhenCarJobIsNull_ThrowsArgumentNullException()
         {
-            _washService = new HandWaxAndShineService(_loggerMock!.Object);
-            var exception = await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await _washService.HandWaxAndShineAsync(null!));
+            _washService = new InteriorCleanService(_loggerMock!.Object);
+            var exception = await Assert.ThrowsExceptionAsync<ArgumentNullException>(async () => await _washService.CleanInteriorAsync(null!));
             Assert.AreEqual("carJob", exception.ParamName);
         }
 
         [TestMethod]
-        public async Task HandWaxAndShineAsync_WhenCalled_PerformsWashAndLogs()
+        public async Task CleanInteriorAsync_WhenCalled_PerformsWashAndLogs()
         {
             // Arrange
-            _washService = new HandWaxAndShineService(_loggerMock!.Object);
+            _washService = new InteriorCleanService(_loggerMock!.Object);
 
             // Act
-            await _washService.HandWaxAndShineAsync(_carJob!);
+            await _washService.CleanInteriorAsync(_carJob!);
 
             // Assert
             // TODO: This approach could be modified by creating a custom ILogger implementation that records logs to a list and asserting against that.
@@ -59,7 +62,7 @@ namespace CarWashProcessor.UnitTests.Services
                 x => x.Log(
                     LogLevel.Information,
                     It.IsAny<EventId>(),
-                    It.Is<It.IsAnyType>((v, t) => v.ToString()!.Contains("--> Hand waxed and shined for customer " + _carJob!.CustomerId)),
+                    It.Is<It.IsAnyType>((v, t) => v.ToString()!.Contains("--> Interior has been cleaned for customer " + _carJob!.CustomerId)),
                     It.IsAny<Exception>(),
                     It.IsAny<Func<It.IsAnyType, Exception?, string>>()),
                 Times.Once);
