+++ CarWashProcessor/Application/Strategies/Addons/InteriorCleanService.cs
@@ -1,16 +1,16 @@
-// Copyright (c) 2025 Car Wash Processor, All Rights Reserved
+// Copyright (c) 2025 Car Wash Processor, All Rights Reserved.
 
-using CarWashProcessor.Models;
+using CarWashProcessor.Application.Abstractions.Registration;   // For AddonTypeAttribute
+using CarWashProcessor.Domain.Abstractions.Services;            // For IAddonServiceStrategy
+using CarWashProcessor.Models;                                  // For CarJob, EServiceAddon    
 
-namespace CarWashProcessor.Services;
+namespace CarWashProcessor.Application.Strategies.Addons;
 
 /// <summary>
 /// Service responsible for interior cleaning as an addon service.
 /// </summary>
-/// <remarks>
-/// TODO: Each addon service has identical structure. Depending on requirements (execution order doesn't matter or it does) this is either a strategy or a command / step.
-/// </remarks>
-public class InteriorCleanService
+[AddonType(EServiceAddon.InteriorClean)]
+public class InteriorCleanService : IAddonServiceStrategy
 {
     /// <summary>
     /// Logger instance for logging information related to the InteriorCleanService.
@@ -28,7 +28,7 @@ public class InteriorCleanService
     /// </exception>
     public InteriorCleanService(ILogger<InteriorCleanService> logger)
 	{
-        // Defensive programming, validate input parameters in constructor, or fast fail, followed by assignments.
+        // Defensive programming.
         ArgumentNullException.ThrowIfNull(logger, nameof(logger));
 
         // Set services
@@ -47,21 +47,18 @@ public class InteriorCleanService
     /// <exception cref="ArgumentNullException">
     /// Thrown if the <paramref name="carJob"/> parameter is null.
     /// </exception>
-	public async Task CleanInteriorAsync(CarJob carJob)
-	{
-        // Defensive programming. Validate input parameters on public methods.
+	public async Task CleanInteriorAsync(CarJob carJob) => await PerformAddonAsync(carJob);
+
+    /// <inheritdoc />
+    public async Task PerformAddonAsync(CarJob carJob, CancellationToken cancellationToken = default)
+    {
+        // Defensive programming.
         ArgumentNullException.ThrowIfNull(carJob, nameof(carJob));
 
         // Wait a second (simulating addon-specific work).
-        await Task.Delay(TimeSpan.FromSeconds(1));
+        await Task.Delay(TimeSpan.FromSeconds(1), cancellationToken);
 
         // Log information
-        /* TODO: Evaluate CA1848: Use LoggerMessage.Define to pre-define logging messages for better performance. 
-         * This is a suggestion from code analysis, but for simplicity and readability in this example, we are 
-         * using the straightforward approach as it can be argued this does not invalidate the requirement: 
-         * "The system is reasonably performant". */
-#pragma warning disable CA1848 // Use the LoggerMessage delegates
         _logger.LogInformation("--> Interior has been cleaned for customer {CustomerId}!", carJob.CustomerId);
-#pragma warning restore CA1848 // Use the LoggerMessage delegates
     }
 }
