+++ CarWashProcessor/Application/Strategies/Addons/TireShineService.cs
@@ -1,0 +1,64 @@
+// Copyright (c) 2025 Car Wash Processor, All Rights Reserved.
+
+using CarWashProcessor.Application.Abstractions.Registration;   // For AddonTypeAttribute
+using CarWashProcessor.Domain.Abstractions.Services;            // For IAddonServiceStrategy
+using CarWashProcessor.Models;                                  // For CarJob, EServiceAddon
+
+namespace CarWashProcessor.Application.Strategies.Addons;
+
+/// <summary>
+/// Service responsible for shining tires as an addon service.
+/// </summary>
+[AddonType(EServiceAddon.TireShine)]
+public class TireShineService : IAddonServiceStrategy
+{
+    /// <summary>
+    /// Logger instance for logging information related to the TireShineService.
+    /// </summary>
+    private readonly ILogger<TireShineService> _logger;
+
+    /// <summary>
+    /// Constructor for TireShineService.
+    /// </summary>
+    /// <param name="logger">
+    /// Logger instance for logging information related to the TireShineService.
+    /// </param>
+    /// <exception cref="ArgumentNullException">
+    /// Thrown if the <paramref name="logger"/> parameter is null.
+    /// </exception>
+	public TireShineService(ILogger<TireShineService> logger)
+	{
+        // Defensive programming.
+        ArgumentNullException.ThrowIfNull(logger, nameof(logger));
+
+        // Set services
+        _logger = logger;
+    }
+
+    /// <inheritdoc />
+    public async Task PerformAddonAsync(CarJob carJob, CancellationToken cancellationToken = default)
+    {
+        // Defensive programming. Validate input parameters on public methods.
+        ArgumentNullException.ThrowIfNull(carJob, nameof(carJob));
+
+        // Wait a second (simulating addon-specific work).
+        await Task.Delay(TimeSpan.FromSeconds(1));
+
+        // Log information
+        _logger.LogInformation("--> Tires have been shined for customer {CustomerId}!", carJob.CustomerId);
+    }
+
+    /// <summary>
+    /// Shines the tires for the given car job.
+    /// </summary>
+    /// <param name="carJob">
+    /// The car job for which the tires are to be shined.
+    /// </param>
+    /// <returns>
+    /// A task representing the asynchronous operation of shining the tires.
+    /// </returns>
+    /// <exception cref="ArgumentNullException">
+    /// Thrown if the <paramref name="carJob"/> parameter is null.
+    /// </exception>
+    public async Task ShineTiresAsync(CarJob carJob) => await PerformAddonAsync(carJob);
+}
